Simplify code (remove redundant code): FFmpegVideoSource and FFmpegAudioSource are base class of FFmpegCameraSource, FFmpegFileSource, FFmpegMicrophoneSource and FFmpegScreenSource. Update README and add webrtc.html file in FFmpegFileAndDeviceTest project.
