+++ src/FFmpegVideoSource.cs
@@ -1,0 +1,181 @@
+﻿using FFmpeg.AutoGen;
+using Microsoft.Extensions.Logging;
+using SIPSorceryMedia.Abstractions;
+using System;
+using System.Collections.Generic;
+using System.Runtime.InteropServices;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace SIPSorceryMedia.FFmpeg
+{
+    public class FFmpegVideoSource: IVideoSource, IDisposable
+    {
+        private static ILogger logger = SIPSorcery.LogFactory.CreateLogger<FFmpegVideoSource>();
+
+        internal static List<VideoFormat> _supportedVideoFormats = new List<VideoFormat>
+        {
+            new VideoFormat(VideoCodecsEnum.VP8, Helper.VP8_FORMATID, Helper.VIDEO_SAMPLING_RATE),
+            new VideoFormat(VideoCodecsEnum.H264, Helper.H264_FORMATID, Helper.VIDEO_SAMPLING_RATE)
+        };
+
+
+        internal bool _isStarted;
+        internal bool _isPaused;
+        internal bool _isClosed;
+
+        internal FFmpegVideoDecoder ? _videoDecoder;
+
+        internal VideoFrameConverter? _videoFrameConverter = null;
+
+        internal FFmpegVideoEncoder _videoEncoder;
+        internal bool _forceKeyFrame;
+
+        internal MediaFormatManager<VideoFormat> _videoFormatManager;
+
+        public event EncodedSampleDelegate? OnVideoSourceEncodedSample;
+#pragma warning disable CS0067
+        public event RawVideoSampleDelegate? OnVideoSourceRawSample;
+        public event SourceErrorDelegate? OnVideoSourceError;
+#pragma warning restore CS0067
+        public event Action? OnEndOfFile;
+
+        public FFmpegVideoSource()
+        {
+            _videoFormatManager = new MediaFormatManager<VideoFormat>(_supportedVideoFormats);
+            _videoEncoder = new FFmpegVideoEncoder();
+        }
+
+        public unsafe void CreateVideoDecoder(String path, AVInputFormat* avInputFormat, bool repeat = false, bool isCamera = false)
+        {
+            _videoDecoder = new FFmpegVideoDecoder(path, avInputFormat, false, isCamera);
+            _videoDecoder.OnVideoFrame += FileSourceDecoder_OnVideoFrame;
+
+            _videoDecoder.OnEndOfFile += () =>
+            {
+                logger.LogDebug($"File source decode complete for {path}.");
+                OnEndOfFile?.Invoke();
+                _videoDecoder.Dispose();
+            };
+        }
+
+        public void InitialiseDecoder(Dictionary<string, string>? decoderOptions = null)
+        {
+            _videoDecoder?.InitialiseSource(decoderOptions);
+        }
+
+
+        public bool IsPaused() => _isPaused;
+
+        public List<VideoFormat> GetVideoSourceFormats()
+        {
+            return _videoFormatManager.GetSourceFormats();
+        }
+
+        public void SetVideoSourceFormat(VideoFormat videoFormat)
+        {
+            logger.LogDebug($"Setting video source format to {videoFormat.FormatID}:{videoFormat.Codec} {videoFormat.ClockRate}.");
+            _videoFormatManager.SetSelectedFormat(videoFormat);
+        }
+        public void RestrictFormats(Func<VideoFormat, bool> filter)
+        {
+            _videoFormatManager.RestrictFormats(filter);
+        }
+
+        public void ForceKeyFrame() => _forceKeyFrame = true;
+        public void ExternalVideoSourceRawSample(uint durationMilliseconds, int width, int height, byte[] sample, VideoPixelFormatsEnum pixelFormat) => throw new NotImplementedException();
+        public bool HasEncodedVideoSubscribers() => OnVideoSourceEncodedSample != null;
+        public bool IsVideoSourcePaused() => _isPaused;
+        public Task StartVideo() => Start();
+        public Task PauseVideo() => Pause();
+        public Task ResumeVideo() => Resume();
+        public Task CloseVideo() => Close();
+
+        private unsafe void FileSourceDecoder_OnVideoFrame(ref AVFrame frame)
+        {
+            if (OnVideoSourceEncodedSample != null)
+            {
+                int frameRate = (int)_videoDecoder.VideoAverageFrameRate;
+                frameRate = (frameRate <= 0) ? Helper.DEFAULT_FRAME_RATE : frameRate;
+                uint timestampDuration = (uint)(Helper.VIDEO_SAMPLING_RATE / frameRate);
+
+                var width = frame.width;
+                var height = frame.height;
+
+                if (_videoFrameConverter == null ||
+                    _videoFrameConverter.SourceWidth != width ||
+                    _videoFrameConverter.SourceHeight != height)
+                {
+                    _videoFrameConverter = new VideoFrameConverter(
+                        width, height,
+                        (AVPixelFormat)frame.format,
+                        width, height,
+                        AVPixelFormat.AV_PIX_FMT_YUV420P);
+                }
+
+                frameRate = (frameRate <= 0) ? Helper.DEFAULT_FRAME_RATE : frameRate;
+                byte[] sample = _videoFrameConverter.ConvertFrame(ref frame);
+
+                AVCodecID aVCodecId = FFmpegConvert.GetAVCodecID(_videoFormatManager.SelectedFormat.Codec);
+                byte[]? encodedSample = _videoEncoder.Encode(aVCodecId, sample, frame.width, frame.height, frameRate, _forceKeyFrame);
+
+                if (encodedSample != null)
+                {
+                    // Note the event handler can be removed while the encoding is in progress.
+                    OnVideoSourceEncodedSample?.Invoke(timestampDuration, encodedSample);
+
+                    _forceKeyFrame = false;
+                }
+            }
+        }
+
+        public Task Start()
+        {
+            if (!_isStarted)
+            {
+                _isStarted = true;
+                _videoDecoder.StartDecode();
+            }
+
+            return Task.CompletedTask;
+        }
+
+        public async Task Close()
+        {
+            if (!_isClosed)
+            {
+                _isClosed = true;
+                await _videoDecoder.Close();
+                Dispose();
+            }
+        }
+
+        public Task Pause()
+        {
+            if (!_isPaused)
+            {
+                _isPaused = true;
+                _videoDecoder.Pause();
+            }
+
+            return Task.CompletedTask;
+        }
+
+        public async Task Resume()
+        {
+            if (_isPaused && !_isClosed)
+            {
+                _isPaused = false;
+                await _videoDecoder.Resume();
+            }
+        }
+
+        public void Dispose()
+        {
+            _videoDecoder?.Dispose();
+
+            _videoEncoder?.Dispose();
+        }
+
+    }
+}
